Flytt barnDataKeySpørsmål til søknadKontrakt
