+++ src/common/typer/kontrakt/søknadKontrakt.ts
@@ -1,7 +1,5 @@
 import { ESvar } from '@navikt/familie-form-elements';
 
-import { barnDataKeySpørsmål } from '../../../frontend/typer/barn';
-
 import { ISøknadKontraktDokumentasjon } from './dokumentasjon';
 import {
     ERegistrertBostedType,
@@ -65,6 +63,31 @@ export interface ISøknadKontraktSøker {
     idNummer: ISøknadsfelt<IIdNummerIKontraktFormat>[];
     andreUtbetalinger: ISøknadsfelt<ESvar> | null;
     adresseISøkeperiode: ISøknadsfelt<string> | null;
+}
+
+export enum barnDataKeySpørsmål {
+    erFosterbarn = 'erFosterbarn',
+    erAdoptert = 'erAdoptert',
+    erAsylsøker = 'erAsylsøker',
+    kontantstøtteFraAnnetEøsland = 'kontantstøtteFraAnnetEøsland',
+    pågåendeSøknadFraAnnetEøsLand = 'pågåendeSøknadFraAnnetEøsLand',
+    pågåendeSøknadHvilketLand = 'pågåendeSøknadHvilketLand',
+    mottarEllerMottokEøsKontantstøtte = 'mottarEllerMottokEøsKontantstøtte',
+    harBarnehageplass = 'harBarnehageplass',
+    andreForelderErDød = 'andreForelderErDød',
+    oppholderSegIInstitusjon = 'oppholderSegIInstitusjon',
+    utbetaltForeldrepengerEllerEngangsstønad = 'utbetaltForeldrepengerEllerEngangsstønad',
+    boddMindreEnn12MndINorge = 'boddMindreEnn12MndINorge',
+    planleggerÅBoINorge12Mnd = 'planleggerÅBoINorge12Mnd',
+    borFastMedSøker = 'borFastMedSøker',
+    sammeForelderSomAnnetBarnMedId = 'sammeForelderSomAnnetBarnMedId',
+    søkersSlektsforhold = 'søkersSlektsforhold',
+    søkersSlektsforholdSpesifisering = 'søkersSlektsforholdSpesifisering',
+    borMedAndreForelder = 'borMedAndreForelder',
+    borMedOmsorgsperson = 'borMedOmsorgsperson',
+    adresse = 'adresse',
+    søkerDeltKontantstøtte = 'søkerDeltKontantstøtte',
+    foreldreBorSammen = 'foreldreBorSammen',
 }
 
 export interface ISøknadIKontraktBarn {
