+++ src/frontend/components/SøknadsSteg/EøsSteg/Barn/SamletIdNummerForBarn.tsx
@@ -6,11 +6,13 @@ import type { Felt, ISkjema } from '@navikt/familie-skjema';
 
 import { useAppContext } from '../../../../context/AppContext';
 import { useEøsContext } from '../../../../context/EøsContext';
-import { barnDataKeySpørsmål, IBarnMedISøknad } from '../../../../typer/barn';
+import { IBarnMedISøknad } from '../../../../typer/barn';
 import { IEøsForBarnFeltTyper } from '../../../../typer/skjema';
 import { skalSpørreOmIdNummerForPågåendeSøknadEøsLand } from '../../../../utils/barn';
 import { IdNummer } from '../IdNummer';
 import { idNummerLandMedPeriodeType, PeriodeType } from '../idnummerUtils';
+
+import { barnDataKeySpørsmål } from '../../../../../common/typer/kontrakt/søknadKontrakt';
 
 const IdNummerForBarn: React.FC<{
     landAlphaCode: Alpha3Code | '';
