+++ src/frontend/components/SøknadsSteg/VelgBarn/VelgBarn.test.tsx
@@ -7,13 +7,15 @@ import { mockDeep } from 'vitest-mock-extended';
 
 import { ESvar } from '@navikt/familie-form-elements';
 
-import { barnDataKeySpørsmål, IBarnMedISøknad } from '../../../typer/barn';
+import { IBarnMedISøknad } from '../../../typer/barn';
 import { IBarnRespons } from '../../../typer/person';
 import { ISøknad } from '../../../typer/søknad';
 import { mekkGyldigSøker, spyOnUseApp, TestProvidere } from '../../../utils/testing';
 import { OmBarnaDineSpørsmålId } from '../OmBarnaDine/spørsmål';
 
 import VelgBarn from './VelgBarn';
+
+import { barnDataKeySpørsmål } from '../../../../common/typer/kontrakt/søknadKontrakt';
 
 vi.mock('@navikt/fnrvalidator');
