+++ src/frontend/utils/mappingTilKontrakt/andreForelder.ts
@@ -1,8 +1,8 @@
 import { ESvar } from '@navikt/familie-form-elements';
 
-import { barnDataKeySpørsmål, IAndreForelder, IBarnMedISøknad } from '../../typer/barn';
+import { IAndreForelder, IBarnMedISøknad } from '../../typer/barn';
 import { TilRestLocaleRecord } from '../../../common/typer/kontrakt/generelle';
-import { IAndreForelderIKontraktFormat } from '../../../common/typer/kontrakt/søknadKontrakt';
+import { barnDataKeySpørsmål, IAndreForelderIKontraktFormat } from '../../../common/typer/kontrakt/søknadKontrakt';
 import { PersonType } from '../../typer/personType';
 import { ITekstinnhold } from '../../typer/sanity/tekstInnhold';
 import { landkodeTilSpråk } from '../språk';
